parameterize the cleanup since we could either want to reuse, or else previous run failed due to connection timeout

+++ product/create-tags.Jenkinsfile
@@ -4,6 +4,7 @@
 // def MIDSTM_BRANCH = "crw-2.5-rhel-8" // redhat-developer/codeready-workspaces branch to use as source of the new branches
 // def FUTURE_BRANCH = "crw-2.6-rhel-8" // redhat-developer/codeready-workspaces branch to use as source of the scripts run
 // def CRW_TAG = "2.5.0" // tag to create
+// def CLEAN = "false" or "true" // clean /tmp checkouts before proceeding to tag?
 
 def buildNode = "rhel7-releng||rhel7-32gb||rhel7-16gb||rhel7-8gb" // node label
 timeout(120) {
@@ -13,7 +14,9 @@ timeout(120) {
                 sh('curl -sSLO https://raw.githubusercontent.com/redhat-developer/codeready-workspaces/'+ FUTURE_BRANCH + '/product/util.groovy')
                 def util = load "${WORKSPACE}/util.groovy"
                 cleanWs()
-                sh('rm -fr /tmp/tmp-checkouts/')
+                if (CLEAN.trim().equals("true")) {
+                    sh('rm -fr /tmp/tmp-checkouts/')
+                }
                 sh('curl -sSLO https://raw.githubusercontent.com/redhat-developer/codeready-workspaces/'+ FUTURE_BRANCH + '/product/tagRelease.sh && chmod +x tagRelease.sh')
                 sh('curl -sSLO https://raw.githubusercontent.com/redhat-developer/codeready-workspaces/'+ FUTURE_BRANCH + '/product/containerExtract.sh && chmod +x containerExtract.sh')
                 withCredentials([string(credentialsId:'devstudio-release.token', variable: 'GITHUB_TOKEN'), file(credentialsId: 'crw-build.keytab', variable: 'CRW_KEYTAB')]) {
